Extract analysis utils. (#2067)
